requirement.txt
PyYAML==6.0
PyGithub==1.59.0
logging-json==0.4.0
